// schemas/relational-db/src/ddl/poll/revision/structure/PollFactorPositionRevision.ts
import {
	Column,
	DbNumber,
	DbString,
	Entity,
	GeneratedValue,
	Id,
	JoinColumn,
	ManyToOne,
	OneToMany,
	Table
}                       from '@airport/air-control'
import {CascadeType}    from '@airport/ground-control'
import {
	PollFactorPositionRevision_Axis,
	PollFactorPositionRevision_Dir,
	PollFactorPositionRevision_Id
}                       from '../../../../types/poll/revision/structure/PollFactorPositionRevision'
import {FactorPosition} from '../../../factor/position/FactorPosition'
import {ImmutableRow} from '../../../infrastructure/row/ImmutableRow'
import {PollRevision} from '../PollRevision'

@Entity()
@Table({name: 'POLL_FACTOR_POSITION_REVISIONS'})
export class PollFactorPositionRevision
	extends ImmutableRow {

	@GeneratedValue()
	@Id()
	@Column({name: 'POLL_FACTOR_POSITION_REVISION_ID'})
	id: PollFactorPositionRevision_Id

	@Column({name: 'FACTOR_COORDINATE_AXIS', nullable: false})
	@DbString()
	axis: PollFactorPositionRevision_Axis

	@Column({name: 'POSITION_ORIENTATION', nullable: false})
	@DbNumber()
	dir: PollFactorPositionRevision_Dir

	@ManyToOne()
	@JoinColumn({name: 'POLL_REVISION_ID'})
	pollRevision: PollRevision

	/**
	 * This is done at Factor + Position level (vs FactorRevision & PositionRevision)
	 * because the underlying relationship between Factor and Position does not change
	 * with revisions.
	 *
	 * Essentially this just says which positions belong to which factors.  Which
	 * revisions is determined at PollRevision level (not PollFactorPositionRevision)
	 */
	@ManyToOne()
	@JoinColumn({name: 'FACTOR_POSITION_ID', nullable: false})
	factorPosition: FactorPosition

	@ManyToOne()
	@JoinColumn({
		name: 'PARENT_POLL_FACTOR_POSITION_REVISION_ID',
		referencedColumnName: 'POLL_FACTOR_POSITION_REVISION_ID'
	})
	parent: PollFactorPositionRevision

	@OneToMany({cascade: CascadeType.ALL, mappedBy: 'parent'})
	children: PollFactorPositionRevision[]
}